+++ create_test/unrolled_small_matmul_cache_optimal.bpf.c
@@ -5,9 +5,9 @@
 #define IB 4
 int main() {
     void *mat_map_1 = MAP_BY_FD(0), *mat_map_2 = MAP_BY_FD(1), *mat_map_res = MAP_BY_FD(2), *result;
-    float mat_1[MAT_SIZE], mat_2[MAT_SIZE], mat_res[MAT_SIZE];
+    int mat_1[MAT_SIZE], mat_2[MAT_SIZE], mat_res[MAT_SIZE];
     long i, j, k, ii, kk;
-    float acc00, acc01, acc10, acc11;
+    int acc00, acc01, acc10, acc11;
 
     i = 0;
     result = bpf_map_lookup_elem(mat_map_1, &i);
